Patch from Trent Piepho:
This adds a "-x" option to decode-dimms.pl, which lets one supply a list of
file names to read SPD data from.  It can parse various hexdump formats, such
as the output from i2cdump, eeprog, and the util-linux and Busybox hexdump
progams run on a sysfs eeprom file.

Useful for decoding SPD data that you cut and pasted from a manufacturer's
website or from a DIMM installed on an embedded system that does not have
perl/etc, but does have a serial console with busybox.


git-svn-id: http://lm-sensors.org/svn/i2c-tools/trunk@5157 7894878c-1315-0410-8ee3-d5d059ff63e0