decode-vaio: Scan more i2c buses
While the laptop I originally developed decode-vaio on, only had 5 i2c
buses, there could be more on other models, and there are definitely
more on the system I use to test the script (using i2c-stub) these
days. So look for the VAIO EEPROM on up to 32 i2c buses to be on the
safe side.

Signed-off-by: Jean Delvare <jdelvare@suse.de>